Update paket.lock and restore targets
I am very unclear on whether these changes should
be committed or not. The paket restore resulted
in these changes being made. I am not sure if this
is due to a newer version of paket being used, or
whether it is because I am running a later version
of dotnet core!?